+++ udata/core/topic/models.py
@@ -4,10 +4,8 @@ from mongoengine.signals import post_delete, post_save
 
 from udata.api_fields import field
 from udata.core.activity.models import Auditable
-from udata.core.dataset.models import Dataset
 from udata.core.linkable import Linkable
 from udata.core.owned import Owned, OwnedQuerySet
-from udata.core.reuse.models import Reuse
 from udata.models import SpatialCoverage, db
 from udata.search import reindex
 from udata.tasks import as_task_param
@@ -20,7 +18,7 @@ class TopicElement(Auditable, db.Document):
     description = field(db.StringField(required=False))
     tags = field(db.ListField(db.StringField()))
     extras = field(db.ExtrasField())
-    element = field(db.GenericReferenceField(choices=[Dataset, Reuse]))
+    element = field(db.GenericReferenceField(choices=["Dataset", "Reuse", "Dataservice"]))
     # Made optional to allow proper form handling with commit=False
     topic = field(db.ReferenceField("Topic", required=False))